# frontend/openapi.json
{
    "openapi": "3.0.0",
    "components": {
...
        "schemas": {
...
            "ErrorResponse": {
                "properties": { "error": { "type": "string" }, "message": { "type": "string" } },
                "required": ["error", "message"],
                "type": "object",
                "additionalProperties": false
            },
...
            "SuccessResponse": {
                "properties": { "message": { "type": "string" } },
                "required": ["message"],
                "type": "object",
                "additionalProperties": false
            },
...
            "RecipientWithoutPassword": {
                "properties": {
                    "id": { "type": "string" },
                    "name": { "type": "string" },
                    "email": { "type": "string" },
                    "phone": { "type": "string" },
                    "address": { "type": "string" },
                    "fourHandCareNeeded": { "type": "boolean" },
                    "caregiverNote": { "type": "string", "nullable": true }
                },
                "required": ["id", "name", "email", "phone", "address", "fourHandCareNeeded", "caregiverNote"],
                "type": "object",
                "additionalProperties": false
            },
            "CreateRecipientRequest": {
                "properties": {
                    "caregiverNote": { "type": "string" },
                    "fourHandCareNeeded": { "type": "boolean" },
                    "address": { "type": "string" },
                    "password": { "type": "string" },
                    "phone": { "type": "string" },
                    "email": { "type": "string" },
                    "name": { "type": "string" }
                },
                "required": ["caregiverNote", "fourHandCareNeeded", "address", "password", "phone", "email", "name"],
                "type": "object"
            },
            "UpdateRecipientRequest": {
                "properties": {
                    "caregiverNote": { "type": "string" },
                    "fourHandCareNeeded": { "type": "boolean" },
                    "address": { "type": "string" },
                    "phone": { "type": "string" },
                    "email": { "type": "string" },
                    "name": { "type": "string" }
                },
                "required": ["caregiverNote", "fourHandCareNeeded", "address", "phone", "email", "name"],
                "type": "object"
            },
            "UpdateRecipientPasswordRequest": {
                "properties": { "currentPassword": { "type": "string" }, "newPassword": { "type": "string" } },
                "required": ["currentPassword", "newPassword"],
                "type": "object",
                "additionalProperties": false
            },
...
        },
        "securitySchemes": { "jwt": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } }
    },
    "info": { "title": "activity-log-app-backend", "version": "0.0.0", "license": { "name": "ISC" }, "contact": {} },
    "paths": {
...
        "/recipients": {
            "get": {
                "operationId": "GetRecipients",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "items": { "$ref": "#/components/schemas/RecipientWithoutPassword" },
                                            "type": "array"
                                        },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Recipients"],
                "security": [{ "jwt": [] }],
                "parameters": []
            },
            "post": {
                "operationId": "CreateRecipient",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "$ref": "#/components/schemas/RecipientWithoutPassword" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing fields",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Recipients"],
                "security": [{ "jwt": [] }],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": { "schema": { "$ref": "#/components/schemas/CreateRecipientRequest" } }
                    }
                }
            }
        },
        "/recipients/{id}": {
            "get": {
                "operationId": "GetRecipient",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "$ref": "#/components/schemas/RecipientWithoutPassword" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recipient not found",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Recipients"],
                "security": [{ "jwt": [] }],
                "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }]
            },
            "put": {
                "operationId": "UpdateRecipient",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "$ref": "#/components/schemas/SuccessResponse" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing fields",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    },
                    "404": {
                        "description": "Recipient not found",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Recipients"],
                "security": [{ "jwt": [] }],
                "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": { "schema": { "$ref": "#/components/schemas/UpdateRecipientRequest" } }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteRecipient",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "$ref": "#/components/schemas/SuccessResponse" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recipient not found",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Recipients"],
                "security": [{ "jwt": [] }],
                "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }]
            }
        },
        "/recipients/{id}/password": {
            "put": {
                "operationId": "UpdateRecipientPassword",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        { "$ref": "#/components/schemas/SuccessResponse" },
                                        { "$ref": "#/components/schemas/ErrorResponse" }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing fields or invalid password",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    },
                    "404": {
                        "description": "Recipient not found",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
                        }
                    }
                },
                "tags": ["Recipients"],
                "security": [{ "jwt": [] }],
                "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "$ref": "#/components/schemas/UpdateRecipientPasswordRequest" }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [{ "url": "/" }]
}